fixed: 스크롤 시 그리드 자동 연장
